s19d4

+++ src/main/java/com/workintech/s19d1/service/ActorService.java
@@ -18,7 +18,7 @@ public interface ActorService {
 
     Actor save(Actor actor);
 
-    Actor delete(int id);
+    Actor delete(long id);
 
     Actor update(ActorRequest actor, int id);
 }
